Cutting the Planet
A fun game that is currently in the works, and you can run it on LEGO-mindstorms-ev3. If you want to improve it, it supports LEGO MINDSTORMS Education EV3, and I'll have a playable version in the future, but it's just a simple foundation for now, so I didn't annotate the version, it will amaze you. I hope you can point out the error and remind me and thank you.